pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "comidinhas-bff"
version = "0.1.0"
description = "Scaffold inicial de um BFF/BFA em Python com FastAPI."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1.0",
    "httpx>=0.28,<1.0",
    "pydantic-settings>=2.6,<3.0",
    "python-multipart>=0.0.20,<1.0",
    "uvicorn[standard]>=0.32,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
